refactor: consolidate user-related interfaces and update imports across components
