+++ src/domain/user/user.entity.ts
@@ -4,3 +4,17 @@ export interface User {
   email: string;
   avatarUrl?: string;
 }
+
+export interface BaseUserData {
+  email: string;
+}
+export interface UserData extends BaseUserData {
+  username?: string;
+  avatarUrl?: string;
+}
+
+export interface MembersData extends BaseUserData {
+  id: number;
+  username?: string;
+  avatarUrl?: string;
+}
