+++ src/presentation/widgets/projects/AddParticipantForm/index.ts
@@ -1,1 +1,1 @@
-export { AddParticipantForm, type BaseUserData } from "./AddParticipantForm";
+export { AddParticipantForm } from "./AddParticipantForm";
