--- src/presentation/widgets/projects/MembersList/MembersList.mock.ts
@@ -1,130 +1,0 @@
-import { BaseUserData } from "../AddParticipantForm/AddParticipantForm";
-
-export interface MembersData extends BaseUserData {
-  id: number;
-  username?: string;
-  avatarUrl?: string;
-}
-
-export const users: MembersData[] = [
-  {
-    id: 1,
-    username: "Alice Johnson",
-    email: "alice.johnson@example.com",
-    avatarUrl: "https://randomuser.me/api/portraits/women/1.jpg",
-  },
-  {
-    id: 2,
-    username: "Bob Smith",
-    email: "bob.smith@example.com",
-    avatarUrl: "https://randomuser.me/api/portraits/men/2.jpg",
-  },
-  {
-    id: 3,
-    username: "Carol Lee",
-    email: "carol.lee@example.com",
-    avatarUrl: "https://randomuser.me/api/portraits/women/3.jpg",
-  },
-  {
-    id: 4,
-    username: "David Kim",
-    email: "david.kim@example.com",
-    avatarUrl: "https://randomuser.me/api/portraits/men/4.jpg",
-  },
-  {
-    id: 5,
-    username: "Eva Brown",
-    email: "eva.brown@example.com",
-    avatarUrl: "https://randomuser.me/api/portraits/women/5.jpg",
-  },
-  {
-    id: 6,
-    username: "Frank Miller",
-    email: "frank.miller@example.com",
-    avatarUrl: "https://randomuser.me/api/portraits/men/6.jpg",
-  },
-  {
-    id: 7,
-    username: "Grace Wilson",
-    email: "grace.wilson@example.com",
-    avatarUrl: "https://randomuser.me/api/portraits/women/7.jpg",
-  },
-  {
-    id: 8,
-    username: "Henry Moore",
-    email: "henry.moore@example.com",
-    avatarUrl: "https://randomuser.me/api/portraits/men/8.jpg",
-  },
-  {
-    id: 9,
-    username: "Ivy Clark",
-    email: "ivy.clark@example.com",
-    avatarUrl: "https://randomuser.me/api/portraits/women/9.jpg",
-  },
-  {
-    id: 10,
-    username: "Jack Turner",
-    email: "jack.turner@example.com",
-    avatarUrl: "https://randomuser.me/api/portraits/men/10.jpg",
-  },
-  {
-    id: 11,
-    username: "Karen Harris",
-    email: "karen.harris@example.com",
-    avatarUrl: "https://randomuser.me/api/portraits/women/11.jpg",
-  },
-  {
-    id: 12,
-    username: "Leo Martin",
-    email: "leo.martin@example.com",
-    avatarUrl: "https://randomuser.me/api/portraits/men/12.jpg",
-  },
-  {
-    id: 13,
-    username: "Mia Thompson",
-    email: "mia.thompson@example.com",
-    avatarUrl: "https://randomuser.me/api/portraits/women/13.jpg",
-  },
-  {
-    id: 14,
-    username: "Noah White",
-    email: "noah.white@example.com",
-    avatarUrl: "https://randomuser.me/api/portraits/men/14.jpg",
-  },
-  {
-    id: 15,
-    username: "Olivia Hall",
-    email: "olivia.hall@example.com",
-    avatarUrl: "https://randomuser.me/api/portraits/women/15.jpg",
-  },
-  {
-    id: 16,
-    username: "Paul Young",
-    email: "paul.young@example.com",
-    avatarUrl: "https://randomuser.me/api/portraits/men/16.jpg",
-  },
-  {
-    id: 17,
-    username: "Quinn King",
-    email: "quinn.king@example.com",
-    avatarUrl: "https://randomuser.me/api/portraits/women/17.jpg",
-  },
-  {
-    id: 18,
-    username: "Ryan Scott",
-    email: "ryan.scott@example.com",
-    avatarUrl: "https://randomuser.me/api/portraits/men/18.jpg",
-  },
-  {
-    id: 19,
-    username: "Sophie Green",
-    email: "sophie.green@example.com",
-    avatarUrl: "https://randomuser.me/api/portraits/women/19.jpg",
-  },
-  {
-    id: 20,
-    username: "Tom Baker",
-    email: "tom.baker@example.com",
-    avatarUrl: "https://randomuser.me/api/portraits/men/20.jpg",
-  },
-];
